Moved numeric extensions to a separate file, added abs_r method.
